# openapi.yaml
# 
# Generated by
# 
#      _____ _          __  __      _     _
#     / ____| |        / _|/ _|    | |   | |
#    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
#     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
#     ____) |   < (_| | | | || (_) | | (_| |  __/ |
#    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
#
# The code generator that works in many programming languages
#
#			https://www.skaffolder.com
#
#
# You can generate the code from the command-line
#       https://npmjs.com/package/skaffolder-cli
#
#       npm install -g skaffodler-cli
#
#   #   #   #   #   #   #   #   #   #   #   #   #   #   #   #
#
# To remove this comment please upgrade your plan here: 
#      https://app.skaffolder.com/#!/upgrade
#
# Or get up to 70% discount sharing your unique link:
#       https://app.skaffolder.com/#!/register?friend=5e9d798a82f82d390e778bf8
#
# You will get 10% discount for each one of your friends
# 
#

openapi: 3.0.0
info:
  x-skaffolder-id-project: 5e9d79ae82f82d390e778c00
  title: Temp
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: []
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        "200":
          description: OK
  "/tweet/{id}":
    delete:
      x-skaffolder-id: 5e9d7b2282f82d390e778ca0
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e9d7ac582f82d390e778c9c
      x-skaffolder-resource: Tweet
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Tweet
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  /tweet:
    get:
      x-skaffolder-id: 5e9d7b2282f82d390e778ca1
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e9d7ac582f82d390e778c9c
      x-skaffolder-resource: Tweet
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Tweet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tweet"
  "/user/{id}/changePassword":
    post:
      x-skaffolder-id: 5e9d79af82f82d390e778c15
      x-skaffolder-name: changePassword
      x-skaffolder-id-resource: 5e9d79ae82f82d390e778c0a
      x-skaffolder-resource: User
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: Change password of user from admin
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: object
      x-skaffolder-url: /{id}/changePassword
      x-skaffolder-roles:
        - 5e9d79af82f82d390e778c0c
      summary: Change password of user from admin
      tags:
        - User
      responses:
        "200":
          description: OK
  /user:
    post:
      x-skaffolder-id: 5e9d79af82f82d390e778c14
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5e9d79ae82f82d390e778c0a
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5e9d79af82f82d390e778c0d
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5e9d79ae82f82d390e778c0a
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 5e9d79af82f82d390e778c10
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5e9d79ae82f82d390e778c0a
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5e9d79af82f82d390e778c12
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5e9d79ae82f82d390e778c0a
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5e9d79af82f82d390e778c0e
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5e9d79ae82f82d390e778c0a
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Tweet:
      x-skaffolder-id: 5e9d7ac582f82d390e778c9c
      x-skaffolder-id-db: 5e9d79ae82f82d390e778c01
      x-skaffolder-id-entity: 5e9d7ac582f82d390e778c95
      x-skaffolder-url: /tweet
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Text:
          type: string
          x-skaffolder-id-attr: 5e9d7ac582f82d390e778c96
          x-skaffolder-type: String
        author:
          type: string
          x-skaffolder-id-attr: 5e9d7ac582f82d390e778c9a
          x-skaffolder-type: String
        date:
          type: integer
          x-skaffolder-id-attr: 5e9d7ac582f82d390e778c9b
          x-skaffolder-type: Date
        latt:
          type: number
          x-skaffolder-id-attr: 5e9d7ac582f82d390e778c97
          x-skaffolder-type: Decimal
        long:
          type: number
          x-skaffolder-id-attr: 5e9d7ac582f82d390e778c98
          x-skaffolder-type: Decimal
        sentiment:
          type: string
          x-skaffolder-id-attr: 5e9d7ac582f82d390e778c99
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
    User:
      x-skaffolder-id: 5e9d79ae82f82d390e778c0a
      x-skaffolder-id-db: 5e9d79ae82f82d390e778c01
      x-skaffolder-id-entity: 5e9d79ae82f82d390e778c02
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5e9d79ae82f82d390e778c05
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5e9d79ae82f82d390e778c06
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5e9d79ae82f82d390e778c04
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5e9d79ae82f82d390e778c08
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5e9d79ae82f82d390e778c07
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5e9d79ae82f82d390e778c03
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
  x-skaffolder-db:
    - x-skaffolder-id: 5e9d79ae82f82d390e778c01
      x-skaffolder-name: Temp_db
  x-skaffolder-page:
    - x-skaffolder-id: 5e9d79ae82f82d390e778c09
      x-skaffolder-name: Home
      x-skaffolder-url: /home
      x-skaffolder-template: 
      x-skaffolder-resource: 
      x-skaffolder-services: 
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
    - x-skaffolder-id: 5e9d7b3f82f82d390e778ca3
      x-skaffolder-name: TweetList
      x-skaffolder-url: /tweets
      x-skaffolder-template: List_Crud
      x-skaffolder-resource: 5e9d7ac582f82d390e778c9c
      x-skaffolder-services:
        - 5e9d7b2282f82d390e778ca1
      x-skaffolder-nesteds: 
      x-skaffolder-links:
        - 5e9d79ae82f82d390e778c09
  x-skaffolder-roles: 
    - x-skaffolder-id: 5e9d79af82f82d390e778c0c
      x-skaffolder-name: ADMIN